[Fix]Class명 변경
