+++ src/test/java/collab/ModifyByIDTest.java
@@ -2,11 +2,10 @@ package collab;
 
 import static org.junit.jupiter.api.Assertions.*;
 import java.util.Arrays;
-import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class ModifyTestByCondition {
+class ModifyByIDTest {
   
   DataBase empDB = new DataBase();
     @BeforeEach
@@ -17,56 +16,56 @@ class ModifyTestByCondition {
     }
     
     @Test
-    void modifyTest_sName_cName() {
+    void modifyTest_returnCheck_Name() {
       Employee targetEmployee = new Employee(Arrays.asList("17112609","FB NTAWR","CL4","010-5645-6122","19860903","PRO"));
-      List<Employee> foundEmployee = empDB.modifyItemByCondition("name", "FB NTAWR", "name", "TEST TEST");
+      Employee foundEmployee = empDB.modifyItemById("17112609", "name", "TEST TEST");
 
-      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.get(0).getEmployeeNumber());
-      assertEquals(targetEmployee.getName(), foundEmployee.get(0).getName());
+      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.getEmployeeNumber());
+      assertEquals(targetEmployee.getName(), foundEmployee.getName());
       assertTrue(empDB.searchItems("name", "FB NTAWR").isEmpty());
       assertTrue(!empDB.searchItems("name", "TEST TEST").isEmpty());
     }
     
     @Test
-    void modifyTest_sName_cCl() {
+    void modifyTest_returnCheck_Cl() {
       Employee targetEmployee = new Employee(Arrays.asList("17112609","FB NTAWR","CL4","010-5645-6122","19860903","PRO"));
-      List<Employee> foundEmployee = empDB.modifyItemByCondition("name", "FB NTAWR", "cl", "CL2");
-
-      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.get(0).getEmployeeNumber());
-      assertEquals(targetEmployee.getCareerLevel(), foundEmployee.get(0).getCareerLevel());
+      Employee foundEmployee = empDB.modifyItemById("17112609", "cl", "CL2");
+      
+      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.getEmployeeNumber());
+      assertEquals(targetEmployee.getCareerLevel(), foundEmployee.getCareerLevel());
       assertTrue(empDB.searchItems("cl", "CL4").isEmpty());
       assertTrue(!empDB.searchItems("cl", "CL2").isEmpty());
     }
     
     @Test
-    void modifyTest_sName_cPhoneNumber() {
+    void modifyTest_returnCheck_PhoneNum() {
       Employee targetEmployee = new Employee(Arrays.asList("17112609","FB NTAWR","CL4","010-5645-6122","19860903","PRO"));
-      List<Employee> foundEmployee = empDB.modifyItemByCondition("name", "FB NTAWR", "phoneNum", "010-5645-0000");
+      Employee foundEmployee = empDB.modifyItemById("17112609", "phoneNum", "010-5645-0000");
 
-      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.get(0).getEmployeeNumber());
-      assertEquals(targetEmployee.getPhoneNumber(), foundEmployee.get(0).getPhoneNumber());
+      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.getEmployeeNumber());
+      assertEquals(targetEmployee.getPhoneNumber(), foundEmployee.getPhoneNumber());
       assertTrue(empDB.searchItems("phoneNum", "010-5645-6122").isEmpty());
       assertTrue(!empDB.searchItems("phoneNum", "010-5645-0000").isEmpty());
     }
     
     @Test
-    void modifyTest_sName_cBirthday() {
+    void modifyTest_returnCheck_Birthday() {
       Employee targetEmployee = new Employee(Arrays.asList("17112609","FB NTAWR","CL4","010-5645-6122","19860903","PRO"));
-      List<Employee> foundEmployee = empDB.modifyItemByCondition("name", "FB NTAWR", "birthday", "19860101");
-
-      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.get(0).getEmployeeNumber());
-      assertEquals(targetEmployee.getBirthday(), foundEmployee.get(0).getBirthday());
+      Employee foundEmployee = empDB.modifyItemById("17112609", "birthday", "19870903");
+      
+      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.getEmployeeNumber());
+      assertEquals(targetEmployee.getBirthday(), foundEmployee.getBirthday());
       assertTrue(empDB.searchItems("birthday", "19860903").isEmpty());
-      assertTrue(!empDB.searchItems("birthday", "19860101").isEmpty());
+      assertTrue(!empDB.searchItems("birthday", "19870903").isEmpty());
     }
     
     @Test
-    void modifyTest_sName_cCerti() {
+    void modifyTest_returnCheck_Certi() {
       Employee targetEmployee = new Employee(Arrays.asList("17112609","FB NTAWR","CL4","010-5645-6122","19860903","PRO"));
-      List<Employee> foundEmployee = empDB.modifyItemByCondition("name", "FB NTAWR", "certi", "EX");
-
-      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.get(0).getEmployeeNumber());
-      assertEquals(targetEmployee.getCerti(), foundEmployee.get(0).getCerti());
+      Employee foundEmployee = empDB.modifyItemById("17112609", "certi", "EX");
+      
+      assertEquals(targetEmployee.getEmployeeNumber(), foundEmployee.getEmployeeNumber());
+      assertEquals(targetEmployee.getCerti(), foundEmployee.getCerti());
       assertTrue(empDB.searchItems("certi", "PRO").isEmpty());
       assertTrue(!empDB.searchItems("certi", "EX").isEmpty());
     }
